Desktop terminal client for the input protocol.
Connects stdin from a terminal on the desktop to MicroPython stdin on the device using the base64 message protocol on the serial connection. This allows convenient use of the REPL on the device.

Output is passed through as-is and is not really usable yet due to https://devforum.play.date/t/logtoconsole-without-a-linebreak/1819 and https://devforum.play.date/t/serial-interface-echo-off-does-not-completely-disable-echo/17149 . For now, look at the terminal on the device, not on the desktop machine.

Needs PySerial (pip3 install pyserial), run like
terminal.py /dev/cu.usbmodemPDU1_Y0035561
with the serial port device as an argument.

As a shortcut, the Console class from mpremote is used to set up the terminal for raw input and output. Tested on macOS, probably works on other POSIX systems, no idea about Windows.