Run MicroPython in a coroutine.
This allows having long-running Python code, such as PewPew games that expect to run the main loop themselves, without blocking the Playdate main loop for too long, as long as the Python code regularly yields (which we can do at least in pew.tick(), possibly even in the VM itself with no cooperation from Python code required).

All MicroPython API calls must be done in the coroutine, because MicroPython checks stack usage and gets confused when the stack pointer is in a completely different place. Therefore introduce a queue (ring buffer) for stdin that is fed from the main thread and drained from the Python thread. (This will be needed anyway when switching to the blocking REPL that wants to call rather than be called to get input.)

Using the "C/C++ Coroutines for Playdate" library from https://github.com/nstbayless/playdate-coroutines/pull/1 .