Yield automatically during execution.
`while True: pass` now works without blocking the UI or tripping the watchdog and can be interrupted with ^C.

Yield only when a frame end (return of the update() function) is due, or will be due within the requested wait time, otherwise keep the VM running to avoid slowing it down.

I hope I am doing this the intended way – the situation is not directly comparable to any other port because the 30 FPS ticks are very infrequent in comparison.