Add _pew.show() function to write to the display.
It takes an object with buffer protocol (typically the bytearray in a pew.Pix) and the width (stride).

Optimized for byte-sized buffers of width 8 because that's the usual case. Other widths and buffers of different element sizes are permitted by the pew API and are both handled on the general but slower path.