Fix the need to modify the Playdate SDK.
It turns out that `LDFLAGS += --specs=nosys.specs` can actually be put into our makefile (after including common.mk) and does not need to be patched into common.mk. I originally assumed that that wouldn't work, because the rule using the variable comes before that definition, but it actually does because the variable is a recursively expanded one (evaluated at use), not a simply expanded one (evaluated at definition).